+++ spec/unit/grammar/matching/test_compile_rule.rb
@@ -4,18 +4,50 @@ module Sexpr
     describe Matching, "compile_rule" do
       include Matching
 
-      it 'keep alternatives unchanged' do
-        compile_rule(:hello, Matcher::Alternative.new([]) ).should be_a(Matcher::Alternative)
+      subject{ compile_rule(:hello, defn) }
+
+      shared_examples_for 'a terminal rule' do
+        it{ should be_a(Matcher::Rule) }
+
+        it 'should have the expected defn' do
+          subject.defn.should be(defn)
+        end
+
+        it 'should have the correct name' do
+          subject.name.should eq(:hello)
+        end
       end
 
-      it 'keep terminals unchanged' do
-        compile_rule(:hello, Matcher::Terminal.new(true) ).should be_a(Matcher::Terminal)
+      shared_examples_for 'a non-terminal rule' do
+        it{ should be_a(Matcher::Rule) }
+
+        it 'should have the expected non-terminal' do
+          subject.defn.should be_a(Matcher::NonTerminal)
+          subject.defn.name.should eq(:hello)
+          subject.defn.defn.should be(defn)
+        end
+
+        it 'should have the correct name' do
+          subject.name.should eq(:hello)
+        end
       end
 
-      it 'keep creates a Rule englobing sequences' do
-        compiled = compile_rule(:hello, Matcher::Sequence.new([]) )
-        compiled.should be_a(Matcher::Rule)
-        compiled.name.should eq(:hello)
+      context 'on Alternative' do
+        let(:defn){ Matcher::Alternative.new([]) }
+
+        it_should_behave_like "a terminal rule"
+      end
+
+      context 'on Terminal' do
+        let(:defn){ Matcher::Terminal.new(true) }
+
+        it_should_behave_like "a terminal rule"
+      end
+
+      context 'on Sequence' do
+        let(:defn){ Matcher::Sequence.new([]) }
+
+        it_should_behave_like "a non-terminal rule"
       end
 
     end